@@ -5,16 +5,15 @@ apply plugin: 'kotlin-kapt'
 apply from: '../quality.gradle'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.shagalalab.qarejet"
         minSdkVersion 19
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
-    kapt { generateStubs = true }
     buildTypes {
         release {
             postprocessing {
@@ -29,37 +28,38 @@ android {
 }
 
 ext {
-    supportLibraryVersion = '26.1.0'
+    appCompatVersion = '1.1.0'
     jUnitVersion = '4.12'
     testRunnerVersion = '1.0.1'
-    espressoVersion = '3.0.1'
+    extJunitVersion = '1.1.1'
+    espressoVersion = '3.2.0'
+    designVersion = "1.0.0"
     rxKotlinVersion = '2.2.0'
-    rxAndroidVersion = '2.0.1'
-    daggerVersion = '2.11'
-    roomVersion = '1.0.0'
+    rxAndroidVersion = '2.1.0'
+    daggerVersion = '2.24'
+    roomVersion = '2.1.0'
     mockitoKotlinVersion = '1.5.0'
     jodaVersion = '2.9.9.1'
     mpAndroidChartVersion = 'v3.0.3'
-    ciceroneVersion = '3.0.0'
+    ciceroneVersion = '5.0.0'
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
     // Support libraries
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlinVersion"
-    implementation "com.android.support:appcompat-v7:$supportLibraryVersion"
-    implementation "com.android.support:design:$supportLibraryVersion"
-    implementation "com.android.support:cardview-v7:$supportLibraryVersion"
-    implementation "com.android.support:recyclerview-v7:$supportLibraryVersion"
+    implementation "androidx.appcompat:appcompat:$appCompatVersion"
+    implementation "com.google.android.material:material:$designVersion"
+    implementation "androidx.cardview:cardview:$designVersion"
+    implementation "androidx.recyclerview:recyclerview:$designVersion"
 
     // RxJava
     implementation "io.reactivex.rxjava2:rxkotlin:$rxKotlinVersion"
     implementation "io.reactivex.rxjava2:rxandroid:$rxAndroidVersion"
 
     // Room
-    implementation "android.arch.persistence.room:runtime:$roomVersion"
-    implementation "android.arch.persistence.room:rxjava2:$roomVersion"
-    kapt "android.arch.persistence.room:compiler:$roomVersion"
+    implementation "androidx.room:room-runtime:$roomVersion"
+    implementation "androidx.room:room-rxjava2:$roomVersion"
+    kapt "androidx.room:room-compiler:$roomVersion"
 
     // Dagger
     implementation "com.google.dagger:dagger:$daggerVersion"
@@ -82,7 +82,8 @@ dependencies {
     testImplementation "junit:junit:$jUnitVersion"
     testImplementation "com.nhaarman:mockito-kotlin:$mockitoKotlinVersion"
     androidTestImplementation "com.android.support.test:runner:$testRunnerVersion"
-    androidTestImplementation "com.android.support.test.espresso:espresso-core:$espressoVersion"
+    androidTestImplementation "androidx.test.ext:junit:$extJunitVersion"
+    androidTestImplementation "androidx.test.espresso:espresso-core:$espressoVersion"
 
     implementation project(path: ':domain')
     implementation project(path: ':data')
