@@ -4,23 +4,22 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 26
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
-    kapt { generateStubs = true }
 }
 
 ext {
     jUnitVersion = '4.12'
-    roomVersion = '1.0.0'
-    rxKotlinVersion = '2.1.0'
-    rxAndroidVersion = '2.0.1'
+    roomVersion = '2.1.0'
+    rxKotlinVersion = '2.2.0'
+    rxAndroidVersion = '2.1.0'
     daggerVersion = '2.11'
     mockitoKotlinVersion = '1.5.0'
 }
@@ -38,9 +37,9 @@ dependencies {
     kapt "com.google.dagger:dagger-compiler:$daggerVersion"
 
     // Room
-    implementation "android.arch.persistence.room:runtime:$roomVersion"
-    implementation "android.arch.persistence.room:rxjava2:$roomVersion"
-    kapt "android.arch.persistence.room:compiler:$roomVersion"
+    implementation "androidx.room:room-runtime:$roomVersion"
+    implementation "androidx.room:room-rxjava2:$roomVersion"
+    kapt "androidx.room:room-compiler:$roomVersion"
 
     implementation project(path: ':domain')
 }
