@@ -4,21 +4,21 @@ import android.animation.ValueAnimator
 import android.app.DatePickerDialog
 import android.app.TimePickerDialog
 import android.os.Bundle
-import android.support.design.widget.Snackbar
-import android.support.v4.widget.TextViewCompat
-import android.support.v4.widget.TextViewCompat.AUTO_SIZE_TEXT_TYPE_UNIFORM
-import android.support.v7.app.AppCompatActivity
 import android.widget.ArrayAdapter
 import android.widget.DatePicker
 import android.widget.TimePicker
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.widget.TextViewCompat
+import androidx.core.widget.TextViewCompat.AUTO_SIZE_TEXT_TYPE_UNIFORM
+import com.google.android.material.snackbar.Snackbar
 import com.shagalalab.qarejet.QarejetApp
 import com.shagalalab.qarejet.R
 import com.shagalalab.qarejet.domain.model.Account
 import com.shagalalab.qarejet.domain.model.Category
 import com.shagalalab.qarejet.ui.widget.CategoryAdapter
-import com.shagalalab.qarejet.ui.widget.picker.DatePickerFragment
 import com.shagalalab.qarejet.ui.widget.numberkeyboard.NumberKeyboardView
+import com.shagalalab.qarejet.ui.widget.picker.DatePickerFragment
 import com.shagalalab.qarejet.ui.widget.picker.TimePickerFragment
 import com.shagalalab.qarejet.util.Constants.TRANSACTION_TYPE_EXPENSE
 import com.shagalalab.qarejet.util.getCurrencySign
