@@ -3,8 +3,8 @@ package com.shagalalab.qarejet.ui.widget.picker
 import android.app.Dialog
 import android.app.TimePickerDialog
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
 import android.text.format.DateFormat
+import androidx.fragment.app.DialogFragment
 import java.util.Calendar
 
 class TimePickerFragment : DialogFragment() {
