@@ -1,10 +1,10 @@
 package com.shagalalab.qarejet.data.db
 
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Delete
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import com.shagalalab.qarejet.data.db.model.AccountDbModel
 import com.shagalalab.qarejet.data.db.model.CategoryDbModel
 import com.shagalalab.qarejet.data.db.model.FullTransactionModel
