@@ -1,6 +1,6 @@
 package com.shagalalab.qarejet.data.db
 
-import android.arch.persistence.room.TypeConverter
+import androidx.room.TypeConverter
 import java.util.*
 
 class DateConverter {
