@@ -12,32 +12,32 @@ class AccountRepositoryImpl constructor(var database: Database) : AccountReposit
 
     override fun getAllAccounts(): Single<List<Account>> {
         return database
-                .accountDao
+                .accountDao()
                 .getAccounts()
                 .map(DbToDomainMapper::mapAccountsList)
     }
 
     override fun getAccount(id: Long): Single<Account> {
         return database
-                .accountDao
+                .accountDao()
                 .getAccount(id)
                 .map(DbToDomainMapper::mapAccount)
     }
 
     override fun addAccount(account: Account): Completable {
-        return Completable.fromAction({
+        return Completable.fromAction {
             database
-                    .accountDao
-                    .insertAccount(DomainToDbMapper.mapAccount(account))
-        })
+                .accountDao()
+                .insertAccount(DomainToDbMapper.mapAccount(account))
+        }
     }
 
     override fun addAccounts(accounts: List<Account>): Completable {
-        return Completable.fromAction({
+        return Completable.fromAction {
             database
-                    .accountDao
-                    .insertAccounts(DomainToDbMapper.mapAccountsList(accounts))
-        })
+                .accountDao()
+                .insertAccounts(DomainToDbMapper.mapAccountsList(accounts))
+        }
     }
 
 }
